refactor: 예약 컴포넌트에서 PENDING 상태 처리 제거 및 예약 로직 간소화
