+++ src/components/SeatSelection.tsx
@@ -14,7 +14,6 @@ import {
   UserInfo,
   SeatMapJson,
   SeatMapSection,
-  GetBookingDetail200ResponseDto,
 } from "./type/index";
 // Removed ScrollArea for simpler overflow handling
 import { Tooltip, TooltipTrigger, TooltipContent } from "./ui/tooltip";
@@ -530,7 +529,6 @@ export function SeatSelection({
 
     console.log("Loading performance data for ID:", performanceId);
     setLoading(true);
-    let autoConfirmFailed = false;
     try {
       // Load performance details
       console.log("Fetching performance details...");
@@ -647,7 +645,6 @@ export function SeatSelection({
   const loadSeats = async (scheduleId: number) => {
     console.log("loadSeats called with scheduleId:", scheduleId);
     setLoading(true);
-    let autoConfirmFailed = false;
     try {
       // Clear any previous selection when loading seats for a new schedule
       resetSelection();
@@ -826,7 +823,6 @@ export function SeatSelection({
     }
 
     setLoading(true);
-    let autoConfirmFailed = false;
     try {
       // Build booking seats from selected seat codes and seatMap sections
       if (!seatMap || !Array.isArray(seatMap.sections) || seatMap.sections.length === 0) {
@@ -870,36 +866,7 @@ export function SeatSelection({
       const bookingResponse = await services.booking.createBooking(bookingRequest);
       console.log("예약 응답:", bookingResponse);
 
-      let confirmedBooking = bookingResponse;
-      try {
-        confirmedBooking = await services.booking.adminConfirmBooking(bookingResponse.bookingId);
-        console.log("예약이 즉시 확정되었습니다:", confirmedBooking);
-      } catch (confirmError) {
-        console.error("Failed to auto-confirm booking:", confirmError);
-
-        let verifiedBooking: GetBookingDetail200ResponseDto | null = null;
-        try {
-          verifiedBooking = await services.booking.adminGetBookingDetail(bookingResponse.bookingId);
-        } catch (statusCheckError) {
-          console.error('Failed to verify booking status after confirm error:', statusCheckError);
-        }
-
-        if (verifiedBooking?.status === 'CONFIRMED') {
-          confirmedBooking = verifiedBooking;
-          console.warn('Auto-confirmation succeeded but follow-up fetch failed. Proceeding with confirmed booking.');
-        } else {
-          autoConfirmFailed = true;
-          try {
-            await services.booking.cancelBooking(bookingResponse.bookingId, {
-              reason: 'Auto-confirmation failed',
-            });
-            console.log('Pending booking rolled back after confirm failure');
-          } catch (rollbackError) {
-            console.error('Failed to rollback booking after confirm failure:', rollbackError);
-          }
-          throw confirmError;
-        }
-      }
+      const confirmedBooking = bookingResponse;
 
       // Remove any pending expiration overrides for this booking now that it is confirmed
       try {
@@ -942,11 +909,7 @@ export function SeatSelection({
       onComplete();
     } catch (error) {
       console.error("Failed to create booking:", error);
-      if (autoConfirmFailed) {
-        alert('예약 확정 중 문제가 발생하여 예약이 취소되었습니다. 다시 시도해주세요.');
-      } else {
-        alert("Booking failed. Please try again.");
-      }
+      alert("Booking failed. Please try again.");
 
       // Clear any stale seat selections before prompting the user again
       resetSelection();
