+++ src/components/service/venueService.ts
@@ -1,4 +1,4 @@
-import { apiClient } from "../../libs/api-client";
+import { apiClient } from "./apiService";
 import {
   VenueDto,
 } from "../type/index";
