--- src/libs/api-client.ts
@@ -1,96 +1,0 @@
-import { API_CONFIG } from "../config/api.config";
-
-// 기본 API 클라이언트
-class ApiClient {
-  private baseUrl: string;
-  private timeout: number;
-
-  constructor(
-    baseUrl: string = API_CONFIG.BASE_URL,
-    timeout: number = API_CONFIG.TIMEOUT
-  ) {
-    this.baseUrl = baseUrl;
-    this.timeout = timeout;
-  }
-
-  private getAuthHeaders(): Record<string, string> {
-    const token = localStorage.getItem("authToken");
-    console.log("API 클라이언트 토큰 확인:", token ? "토큰 있음" : "토큰 없음");
-    return token ? { Authorization: `Bearer ${token}` } : {};
-  }
-
-  private async request<T>(
-    endpoint: string,
-    options: RequestInit = {}
-  ): Promise<T> {
-    const url = `${this.baseUrl}${endpoint}`;
-    const headers = {
-      "Content-Type": "application/json",
-      ...this.getAuthHeaders(),
-      ...options.headers,
-    };
-
-    const controller = new AbortController();
-    const timeoutId = setTimeout(() => controller.abort(), this.timeout);
-
-    try {
-      const response = await fetch(url, {
-        ...options,
-        headers,
-        signal: controller.signal,
-      });
-
-      clearTimeout(timeoutId);
-
-      if (!response.ok) {
-        const errorData = await response.json().catch(() => ({}));
-        throw new Error(
-          errorData.message || `HTTP ${response.status}: ${response.statusText}`
-        );
-      }
-
-      return await response.json();
-    } catch (error) {
-      clearTimeout(timeoutId);
-      if (error instanceof Error && error.name === "AbortError") {
-        throw new Error("Request timeout");
-      }
-      throw error;
-    }
-  }
-
-  async get<T>(endpoint: string): Promise<T> {
-    return this.request<T>(endpoint, { method: "GET" });
-  }
-
-  async post<T>(endpoint: string, data?: unknown): Promise<T> {
-    return this.request<T>(endpoint, {
-      method: "POST",
-      body: data ? JSON.stringify(data) : undefined,
-    });
-  }
-
-  async put<T>(endpoint: string, data?: unknown): Promise<T> {
-    return this.request<T>(endpoint, {
-      method: "PUT",
-      body: data ? JSON.stringify(data) : undefined,
-    });
-  }
-
-  async delete<T>(endpoint: string, options?: { data?: unknown }): Promise<T> {
-    return this.request<T>(endpoint, {
-      method: "DELETE",
-      body: options?.data ? JSON.stringify(options.data) : undefined,
-    });
-  }
-
-  async patch<T>(endpoint: string, data?: unknown): Promise<T> {
-    return this.request<T>(endpoint, {
-      method: "PATCH",
-      body: data ? JSON.stringify(data) : undefined,
-    });
-  }
-}
-
-export const apiClient = new ApiClient();
-export default apiClient;
